+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # 1.x branch
+## 1.4 branch
+### 1.4.0
+* updated for `cakephp` 4 and `phpunit` 8.
+
 ## 1.3 branch
+### 1.3.3
+* little fixes;
+* added tests for lower dependencies.
+
+### 1.3.2
+* little fixes.
+
+### 1.3.1
+* requires `me-tools` package for dev;
+* updated for CakePHP 3.7.1 and `php-tools` 1.1.12.
+
 ### 1.3.0
 * removed the `RECAPTCHA_MAILHIDE` constant;
 * updated for CakePHP 3.7.